Implement event handling using anther callback function
Tied to merge the feature of event handling by the mean of using the
same "recevie" function of upper layer. But seems not so proper if we
treat event exactly the same of some kinds of messages.

One draw back of the method is that, the lower need to add a header to
the message so that the upper can distinguish the event from the common
message. That breaks the design;

Also I mute the stdout of test cases